bump test mod to separate file
