+++ src/main.rs
@@ -3,5 +3,7 @@ mod sha256;
 mod wordlist;
 #[macro_use]
 pub mod macros;
+#[cfg(test)]
+mod test_mnemonic;
 
 fn main() {}
